Ashutosh_Kumar_Jha_12340390_CSL303_Lab_8/requirements.txt
psycopg2-binary>=2.9
python-dotenv 
psycopg2-binary
